requirements.txt
Flask==1.1.2
google-cloud-datastore==1.15.3
google-auth==1.23.0
requests==2.24.0
google-cloud-texttospeech==2.2.0
nltk==3.5
scikit-learn>=0.18